# Speed up adaptive reprojection