--- a/reproject/adaptive/core.py
+++ b/reproject/adaptive/core.py
@@ -5,7 +5,8 @@ from __future__ import (absolute_import, division, print_function,
 import numpy as np
 
 from .deforest import map_coordinates
-from ..wcs_utils import efficient_pixel_to_pixel_with_roundtrip
+from ..wcs_utils import (efficient_pixel_to_pixel_with_roundtrip,
+        efficient_pixel_to_pixel)
 
 
 __all__ = ['_reproject_adaptive_2d']
@@ -13,19 +14,26 @@ __all__ = ['_reproject_adaptive_2d']
 
 class CoordinateTransformer:
 
-    def __init__(self, wcs_in, wcs_out):
+    def __init__(self, wcs_in, wcs_out, roundtrip_coords):
         self.wcs_in = wcs_in
         self.wcs_out = wcs_out
+        self.roundtrip_coords = roundtrip_coords
 
     def __call__(self, pixel_out):
         pixel_out = pixel_out[:, :, 0], pixel_out[:, :, 1]
-        pixel_in = efficient_pixel_to_pixel_with_roundtrip(self.wcs_out, self.wcs_in, *pixel_out)
+        if self.roundtrip_coords:
+            pixel_in = efficient_pixel_to_pixel_with_roundtrip(
+                    self.wcs_out, self.wcs_in, *pixel_out)
+        else:
+            pixel_in = efficient_pixel_to_pixel(
+                    self.wcs_out, self.wcs_in, *pixel_out)
         pixel_in = np.array(pixel_in).transpose().swapaxes(0, 1)
         return pixel_in
 
 
 def _reproject_adaptive_2d(array, wcs_in, wcs_out, shape_out, order=1,
-                           return_footprint=True, center_jacobian=False):
+                           return_footprint=True, center_jacobian=False,
+                           roundtrip_coords=True):
     """
     Reproject celestial slices from an n-d array from one WCS to another
     using the DeForest (2004) algorithm [1]_, and assuming all other dimensions
@@ -47,6 +55,9 @@ def _reproject_adaptive_2d(array, wcs_in, wcs_out, shape_out, order=1,
         Whether to return the footprint in addition to the output array.
     center_jacobian : bool
         Whether to compute centered Jacobians
+    roundtrip_coords : bool
+        Whether to veryfiy that coordinate transformations are defined in both
+        directions.
 
     Returns
     -------
@@ -83,7 +94,7 @@ def _reproject_adaptive_2d(array, wcs_in, wcs_out, shape_out, order=1,
     # Create output array
     array_out = np.zeros(shape_out)
 
-    transformer = CoordinateTransformer(wcs_in, wcs_out)
+    transformer = CoordinateTransformer(wcs_in, wcs_out, roundtrip_coords)
     map_coordinates(array_in, array_out, transformer, out_of_range_nan=True,
                     order=order, center_jacobian=center_jacobian)