… docs/celestial.rst
… reproject/adaptive/tests/reference/test_reproject_adaptive_uncentered_jacobian.fits
… reproject/adaptive/tests/test_core.py
… reproject/interpolation/core.py
… reproject/interpolation/high_level.py
… reproject/interpolation/tests/test_core.py